requirements_clean.txt
matplotlib                    3.7.2
numpy                         1.24.1
opencv-python                 4.10.0
opencv-python-headless        4.10.0
pandas                        2.0.3
PyYAML                        6.0.2
torch                         2.4.1+cu124
torchvision                   0.19.1+cu124
tqdm                          4.66.5
ultralytics                   8.3.43
